chore: Car, index.js 폴더명 변경
